requirements.txt
absl-py==1.0.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
gensim==4.2.0
numpy==1.21.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
graphviz==0.19.1
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
joblib==1.1.0
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.0
Markdown==3.3.6
matplotlib==3.4.0
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==2.5
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==23.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
powerlaw==1.5
prompt-toolkit==3.0.38
protobuf==3.19.4
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-louvain==0.15
pytz==2022.1
pyzmq==25.0.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.22.1
scipy==1.7.3
seaborn==0.11.0
six==1.16.0
smart-open==6.3.0
sympy==1.7.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torchinfo==1.6.5
torchviz==0.0.2
tornado==6.2
tqdm==4.48.2
traitlets==5.9.0
typing_extensions==4.1.1
umap-learn==0.5.4
urllib3==1.26.9
wcwidth==0.2.6
Werkzeug==2.0.3
zipp==3.7.0
